Implemented HDR demo app for OpenTK.Backends.Tests. Fixed demo app name display in OpenTK.Backeds.Tests. Make the win32 backend able to report non-integer refresh rates.

+++ src/OpenTK.Platform/Native/Windows/Handles.cs
@@ -140,7 +140,7 @@ namespace OpenTK.Platform.Native.Windows
 
         public Win32.RECT WorkArea { get; set; }
 
-        public int RefreshRate { get; set; }
+        public float RefreshRate { get; set; }
 
         public int BitsPerPixel { get; set; }
